perf: 去掉 v2 的api
