+++ apps/jumpserver/views/swagger.py
@@ -60,20 +60,12 @@ api_info = openapi.Info(
 
 
 def get_swagger_view(version='v1'):
-    from ..urls import api_v1, api_v2
+    from ..urls import api_v1
     from django.urls import path, include
     api_v1_patterns = [
         path('api/v1/', include(api_v1))
     ]
-
-    api_v2_patterns = [
-        path('api/v2/', include(api_v2))
-    ]
-
-    if version == "v2":
-        patterns = api_v2_patterns
-    else:
-        patterns = api_v1_patterns
+    patterns = api_v1_patterns
     schema_view = get_schema_view(
         api_info,
         patterns=patterns,
