--- apps/terminal/api_v2/__init__.py
@@ -1,3 +1,0 @@
-# -*- coding: utf-8 -*-
-#
-from .terminal import *
