+++ apps/terminal/serializers/terminal.py
@@ -3,6 +3,9 @@ from django.utils.translation import ugettext_lazy as _
 
 from common.drf.serializers import BulkModelSerializer, AdaptedBulkListSerializer
 from common.utils import is_uuid
+from users.serializers import ServiceAccountSerializer
+from common.utils import get_request_ip
+
 from ..models import (
     Terminal, Status, Session, Task, CommandStorage, ReplayStorage
 )
@@ -63,9 +66,42 @@ class StatusSerializer(serializers.ModelSerializer):
 
 
 class TaskSerializer(BulkModelSerializer):
-
     class Meta:
         fields = '__all__'
         model = Task
         list_serializer_class = AdaptedBulkListSerializer
         ref_name = 'TerminalTaskSerializer'
+
+
+class TerminalRegistrationSerializer(serializers.ModelSerializer):
+    service_account = ServiceAccountSerializer(read_only=True)
+
+    class Meta:
+        model = Terminal
+        fields = ['name', 'type', 'comment', 'service_account', 'remote_addr']
+        extra_fields = {
+            'remote_addr': {'readonly': True}
+        }
+
+    def is_valid(self, raise_exception=False):
+        valid = super().is_valid(raise_exception=raise_exception)
+        if not valid:
+            return valid
+        data = {'name': self.validated_data.get('name')}
+        kwargs = {'data': data}
+        if self.instance and self.instance.user:
+            kwargs['instance'] = self.instance.user
+        self.service_account = ServiceAccountSerializer(**kwargs)
+        valid = self.service_account.is_valid(raise_exception=True)
+        return valid
+
+    def save(self, **kwargs):
+        instance = super().save(**kwargs)
+        request = self.context.get('request')
+        instance.is_accepted = True
+        if request:
+            instance.remote_addr = get_request_ip(request)
+        sa = self.service_account.save()
+        instance.user = sa
+        instance.save()
+        return instance
