--- apps/terminal/serializers_v2/__init__.py
@@ -1,4 +1,0 @@
-# -*- coding: utf-8 -*-
-#
-
-from .terminal import *
