+++ apps/users/api/__init__.py
@@ -4,4 +4,5 @@
 from .user import *
 from .group import *
 from .profile import *
+from .service_account import *
 from .relation import *
