+++ apps/users/serializers/__init__.py
@@ -1,5 +1,6 @@
 # -*- coding: utf-8 -*-
 #
 from .user import *
+from .profile import *
 from .group import *
 from .realtion import *
