+++ apps/users/urls/api_urls.py
@@ -15,6 +15,7 @@ router = BulkRouter()
 router.register(r'users', api.UserViewSet, 'user')
 router.register(r'groups', api.UserGroupViewSet, 'user-group')
 router.register(r'users-groups-relations', api.UserUserGroupRelationViewSet, 'users-groups-relation')
+router.register(r'service-account-registrations', api.ServiceAccountRegistrationViewSet, 'service-account-registration')
 
 
 urlpatterns = [
